+++ CSS/catalog.scss
@@ -1,1 +1,3 @@
+// Assembling the catalog page
+
 @import './general/general', './general/fonts', './general/scrollbar', './general/header', './catalog/catalog', './general/footer', './general/btn-up';
