+++ CSS/sketches.scss
@@ -1,1 +1,3 @@
+// Assembling the sketches page
+
 @import './general/general', './general/fonts', './general/scrollbar', './general/header', './sketches/first-screen', './sketches/sketches-list', './sketches/comparison', './sketches/comparison-768px', './general/footer', './general/btn-up';
