+++ JS/catalog/_data.js
@@ -1,0 +1,5 @@
+const url = '../../data/db.json';
+
+export async function getProductsDataArray() {
+    return await fetch(url).then(res => res.json());
+}
