+++ _for-dev.doc
@@ -1,0 +1,13 @@
+Здесь я пишу, что еще надо доработать.
+
+1. Удалить body_back и сделать закрытие popup по клику на любую точку, не входящую в popup
+
+2. Сделать везде одинаковую табуляцию
+
+3. Форматы картинок
+
+4. Добавить description всем страницам
+
+5. btnUp
+
+6. Адаптив для шапки
