+++ _readme.doc
@@ -1,0 +1,7 @@
+Здесь собраны некоторые заметки к проекту.
+
+1.0 Структура проекта
+
+2.0 Именование файлов
+
+3.0 Ветки и коммиты git
