--- database/lamps.json
@@ -1,1 +1,0 @@
-[{"id": "01", "name": "AVENIR", "title": "Светильник напольный из&nbsp;кедровой&nbsp;дупляны&nbsp;(01)", "details1": "высота 0,9-1,4м - 145 000 р.", "details2": "высота 1,5-2,1м - 150-210 т.р.", "imgCounter": "7", "attribute": "Светильник"}]
