+++ index.html
@@ -6,13 +6,13 @@
     <title>Кедровый клуб – мебель ручной работы</title>
     <link rel="icon" href="img/graphics/favicone.ico">
     <link rel="stylesheet" href="CSS/main.css">
-    <script defer src="JS/burger-nav.js"></script>
-    <!-- <script defer src="JS/up.js"></script> -->
-    <script defer src="JS/hideHeader.js"></script>
-    <script defer src="JS/scrollbar.js"></script>
-    <script defer src="JS/main.js"></script>
-    <script defer src="JS/homeDark.js"></script>
-    <script defer src="JS/sketches-examples.js"></script>
+    <script defer src="JS/general/burgerNav.js"></script>
+    <script defer src="JS/general/hideHeader.js"></script>
+    <script defer src="JS/general/scrollbar.js"></script>
+    <!-- <script defer src="JS/general/up.js"></script> -->
+    <script defer src="JS/main/catalogSlider.js"></script>
+    <script defer src="JS/main/homeDark.js"></script>
+    <script defer src="JS/main/sketchesExamples.js"></script>
 </head>
 <body>
 <header class="header">
@@ -29,11 +29,11 @@
           Каталог
         </a>
       </li>
-      <li class="header-nav__item">
+      <!-- <li class="header-nav__item">
         <a class="header-nav__link" href="https://cedar-club.ru/process.html">
           Фпрацэссе
         </a>
-      </li>
+      </li> -->
       <li class="header-nav__item">
         <a class="header-nav__link" href="https://cedar-club.ru/sketches.html">
           Эскизы
@@ -128,9 +128,9 @@
             </li>
           </ul>
         </li>
-        <li class="menu__item">
+        <!-- <li class="menu__item">
           <a class="menu__link" href="https://cedar-club.ru/process.html">Фпрацэссе</a>
-        </li>
+        </li> -->
         <li class="menu__item">
           <a class="menu__link" href="https://cedar-club.ru/sketches.html">Эскизы</a>
         </li>
