+++ process.html
@@ -153,6 +153,5 @@
     </nav>
   </div>
 </header>
-<script src="JS/process.js"></script>
 </body>
 </html>
